Commonize delete note after confirm logic
